requirements.txt
Flask~=2.0.1
opencv-python~=4.5.2.54
requests~=2.25.1
Pillow~=8.2.0
torch~=1.9.0
easyocr==1.3.2
numpy==1.20.3
imutils==0.5.4
scikit-image==0.18.1